Moving all contact logic to hook
